requirements.txt
Django
djangorestframework
Pillow
django-cors-headers
djangorestframework-recursive